refactor: move FileModals to 'FileModalScreen'
