+++ src/view/pages/FileModalScreen/MainView.tsx
@@ -4,8 +4,8 @@ import { getStatusBarHeight } from 'react-native-status-bar-height'
 import { Theme, withTheme, Appbar } from 'react-native-paper'
 import { File } from '../../../models'
 import { ScreenView } from '../../design/ScreenView'
-import { ImageModal } from '../../molecules/ImageModal'
-import { VideoModal } from '../../molecules/VideoModal'
+import { ImageModal } from './ImageModal'
+import { VideoModal } from './VideoModal'
 import color from 'color'
 import { FlatList, PanGestureHandler, State } from 'react-native-gesture-handler'
 import Animated from 'react-native-reanimated'
